+++ src/components/collections/Header/utils/getIsMenuHidden.spec.js
@@ -1,22 +1,26 @@
 import { getIsMenuHidden } from './getIsMenuHidden';
 
 describe('getIsMenuHidden', () => {
-  describe('if `menuWidth` is less than or equal to `navigationItemsWidth` plus the minimum logo spacing', () => {
+  describe('if `menuWidth` minus the header padding is less than or equal to `navigationItemsWidth` plus the `logoWidth`', () => {
     it('should return `true`', () => {
       const testCases = [
-        { menuWidth: 0, navigationItemsWidth: 0 },
-        { menuWidth: 250, navigationItemsWidth: 0 },
+        { menuWidth: 0, navigationItemsWidth: 0, logoWidth: 0 },
+        { menuWidth: 250, navigationItemsWidth: 200, logoWidth: 100 },
       ];
 
-      testCases.forEach(({ menuWidth, navigationItemsWidth }) => {
-        const actual = getIsMenuHidden(menuWidth, navigationItemsWidth);
+      testCases.forEach(({ menuWidth, navigationItemsWidth, logoWidth }) => {
+        const actual = getIsMenuHidden(
+          menuWidth,
+          navigationItemsWidth,
+          logoWidth
+        );
 
         expect(actual).toBe(true);
       });
     });
   });
 
-  describe('if `menuWidth` is more than `navigationItemsWidth` plus the minimum logo spacing', () => {
+  describe('if `menuWidth` minus the header padding is more than the `navigationItemsWidth` plus the `logoWidth`', () => {
     it('should return `false`', () => {
       const actual = getIsMenuHidden(1000, 0);